Client-server integration (#7)
* feat(api): :sparkles: added get deployment reference endpoint

* ci: :building_construction: multi-platform build Dockerfile

* fix(deployment): :ambulance: fixed project deployment isolation issue

* ci: :sparkles: updated compose.yml

* refactor(deployment): :wastebasket: clean up and clear out

* feat(api): :iphone: added client package

* ci(api): :truck: added bin path and rename npm module

* fix(deployment): fix https://github.com/CleverseAcademy/cd-compose-deployment/issues/8

* feat(api): :sparkles: added get next deployment cli command

* refactor(client): :recycle: type and config

* fix(client): :wrench: add support for port flag

* feat(client): :stethoscope: deploy: added git option support

* feat(providers): :seedling: store latest config into a file along with previous version at backup

* refactor(config): :recycle: move project name preparation logic to config module instead

+++ api/dto/deploy_new_image.go
@@ -1,7 +1,7 @@
 package dto
 
 type DeployImageDto struct {
-	Priority int8   `json:"p"`
+	Priority uint64 `json:"p"`
 	Ref      string `json:"r"`
 	Service  string `json:"s"`
 	Image    string `json:"i"`
